Move fakedriver to its own module for reuse
