+++ libmachine/provision/utils_test.go
@@ -2,122 +2,17 @@ package provision
 
 import (
 	"fmt"
-	"os/exec"
 	"regexp"
 	"strings"
 	"testing"
 
-	"github.com/docker/machine/drivers"
+	"github.com/docker/machine/drivers/fakedriver"
 	"github.com/docker/machine/libmachine/auth"
-	"github.com/docker/machine/provider"
-	"github.com/docker/machine/state"
 )
-
-type FakeDriver struct{}
-
-func (d FakeDriver) DriverName() string {
-	return "fakedriver"
-}
-
-func (d FakeDriver) AuthorizePort(ports []*drivers.Port) error {
-	return nil
-}
-
-func (d FakeDriver) DeauthorizePort(ports []*drivers.Port) error {
-	return nil
-}
-
-func (d FakeDriver) SetConfigFromFlags(flags drivers.DriverOptions) error {
-	return nil
-}
-
-func (d FakeDriver) GetURL() (string, error) {
-	return "", nil
-}
-
-func (d FakeDriver) GetMachineName() string {
-	return ""
-}
-
-func (d FakeDriver) GetProviderType() provider.ProviderType {
-	return provider.None
-}
-
-func (d FakeDriver) GetIP() (string, error) {
-	return "1.2.3.4", nil
-}
-
-func (d FakeDriver) GetSSHHostname() (string, error) {
-	return "", nil
-}
-
-func (d FakeDriver) GetSSHKeyPath() string {
-	return ""
-}
-
-func (d FakeDriver) GetSSHPort() (int, error) {
-	return 0, nil
-}
-
-func (d FakeDriver) GetSSHUsername() string {
-	return ""
-}
-
-func (d FakeDriver) GetState() (state.State, error) {
-	return state.Running, nil
-}
-
-func (d FakeDriver) PreCreateCheck() error {
-	return nil
-}
-
-func (d FakeDriver) Create() error {
-	return nil
-}
-
-func (d FakeDriver) Remove() error {
-	return nil
-}
-
-func (d FakeDriver) Start() error {
-	return nil
-}
-
-func (d FakeDriver) Stop() error {
-	return nil
-}
-
-func (d FakeDriver) Restart() error {
-	return nil
-}
-
-func (d FakeDriver) Kill() error {
-	return nil
-}
-
-func (d FakeDriver) Upgrade() error {
-	return nil
-}
-
-func (d FakeDriver) StartDocker() error {
-	return nil
-}
-
-func (d FakeDriver) StopDocker() error {
-	return nil
-}
-
-func (d FakeDriver) GetDockerConfigDir() string {
-	return ""
-}
-
-func (d FakeDriver) GetSSHCommand(args ...string) (*exec.Cmd, error) {
-	return &exec.Cmd{}, nil
-}
 
 func TestGenerateDockerConfigBoot2Docker(t *testing.T) {
 	p := &Boot2DockerProvisioner{
-		Driver: FakeDriver{},
+		Driver: &fakedriver.FakeDriver{},
 	}
 	dockerPort := 1234
 	authConfig := auth.AuthOptions{
@@ -155,7 +50,7 @@ func TestGenerateDockerConfigBoot2Docker(t *testing.T) {
 
 func TestMachinePortBoot2Docker(t *testing.T) {
 	p := &Boot2DockerProvisioner{
-		Driver: FakeDriver{},
+		Driver: &fakedriver.FakeDriver{},
 	}
 	dockerPort := 2376
 	bindUrl := fmt.Sprintf("tcp://0.0.0.0:%d", dockerPort)
@@ -187,7 +82,7 @@ func TestMachinePortBoot2Docker(t *testing.T) {
 
 func TestMachineCustomPortBoot2Docker(t *testing.T) {
 	p := &Boot2DockerProvisioner{
-		Driver: FakeDriver{},
+		Driver: &fakedriver.FakeDriver{},
 	}
 	dockerPort := 3376
 	bindUrl := fmt.Sprintf("tcp://0.0.0.0:%d", dockerPort)
